Fix images not being found when using the "compiled" build
- Fix images not being found when using the "compiled" build
- Allow the game to be reset with ESCAPE or R after a game over
- Add docstrings
- Fix type hints

+++ src/enemy.py
@@ -31,7 +31,7 @@ class Enemy(TypedDict):
     color: int
     hp: int
     attack: int
-    speed: float
+    speed: int
     shield: int
     fire_rate: float
     xp: int
